% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/yaml_nocite.R
\name{yaml_nocite}
\alias{yaml_nocite}
\title{Add nocite to Rmarkdown Document}
\usage{
yaml_nocite(file, outfile = file)
}
\arguments{
\item{file}{input filename (Rmd)}

\item{outfile}{output filename (Rmd). Will overwrite \code{file} by default}
}
\value{
Character of name of outfile
}
\description{
Will add all the bibliography entries to the YAML of an Rmarkdown document
}

